Adds 4.21 solution

+++ controllers/blogs.js
@@ -9,7 +9,6 @@ blogsRouter.get("/", async (request, response) => {
 })
 
 blogsRouter.post("/", async (request, response) => {
-  console.log("request.token", request.token)
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: "token missing or invalid" })
@@ -25,11 +24,17 @@ blogsRouter.post("/", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
-})
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
 
-blogsRouter.delete("/:id", async (request, response) => {
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({ error: "only creator can delete blog" })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
